+++ crates/model/src/orders/trailing_stop_market.rs
@@ -490,7 +490,10 @@ impl Order for TrailingStopMarketOrder {
     }
 
     fn apply(&mut self, event: OrderEventAny) -> Result<(), OrderError> {
-        let was_filled = matches!(event, OrderEventAny::Filled(_));
+        let updates_slippage = matches!(
+            event,
+            OrderEventAny::Filled(_) | OrderEventAny::FillVoided(_),
+        );
         let is_order_triggered = matches!(event, OrderEventAny::Triggered(_));
         let ts_event = if is_order_triggered {
             Some(event.ts_event())
@@ -509,7 +512,7 @@ impl Order for TrailingStopMarketOrder {
             self.ts_triggered = ts_event;
         }
 
-        if was_filled && let Some(trigger_price) = self.trigger_price {
+        if updates_slippage && let Some(trigger_price) = self.trigger_price {
             self.core.set_slippage(trigger_price);
         }